+++ backend/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "blog-backend"
-version = "0.3.0"
+version = "0.5.0"
 authors = ["Songday <songday@yeah.net>"]
 edition = "2021"
