+++ backend/src/resource/asset/snippets/blog-frontend-0b390d584a933dd3/asset/show.js
@@ -1,0 +1,9 @@
+export function showNotificationBox() {
+    document.getElementById('notification').style.display = 'block';
+}
+export function hideNotificationBox(event) {
+    let source = event.target || event.srcElement;
+    while (source.id !== 'notification' && source.parentNode)
+        source = source.parentNode;
+    source.style.display = 'none';
+}
