+++ backend/src/resource/page/index.html
@@ -16,7 +16,7 @@
     
     <link rel="stylesheet" href="/asset/index-ec4ef66164c33b9c.css">
 
-<link rel="preload" href="/asset/index-1d729f6eed0aac2e_bg.wasm" as="fetch" type="application/wasm" crossorigin="">
-<link rel="modulepreload" href="/asset/index-1d729f6eed0aac2e.js"></head>
+<link rel="preload" href="/asset/index-f136c4aa496c3185_bg.wasm" as="fetch" type="application/wasm" crossorigin="">
+<link rel="modulepreload" href="/asset/index-f136c4aa496c3185.js"></head>
 <body>
-<script type="module">import init from '/asset/index-1d729f6eed0aac2e.js';init('/asset/index-1d729f6eed0aac2e_bg.wasm');</script></body></html>
+<script type="module">import init from '/asset/index-f136c4aa496c3185.js';init('/asset/index-f136c4aa496c3185_bg.wasm');</script></body></html>
