+++ backend/src/service/asset_list.rs
@@ -3,13 +3,14 @@
 ("bulma.min-82aac43507618108.css", include_bytes!(r#"..\resource\asset/bulma.min-82aac43507618108.css.gz"#), "text/css"),
 ("codemirror.min.css", include_bytes!(r#"..\resource\asset/codemirror.min.css.gz"#), "text/css"),
 ("fontawesome.min-5e9e696c59c57e83.css", include_bytes!(r#"..\resource\asset/fontawesome.min-5e9e696c59c57e83.css.gz"#), "text/css"),
-("index-1d729f6eed0aac2e.js", include_bytes!(r#"..\resource\asset/index-1d729f6eed0aac2e.js.gz"#), "text/javascript"),
-("index-1d729f6eed0aac2e_bg.wasm", include_bytes!(r#"..\resource\asset/index-1d729f6eed0aac2e_bg.wasm.gz"#), "application/wasm"),
 ("index-ec4ef66164c33b9c.css", include_bytes!(r#"..\resource\asset/index-ec4ef66164c33b9c.css.gz"#), "text/css"),
+("index-f136c4aa496c3185.js", include_bytes!(r#"..\resource\asset/index-f136c4aa496c3185.js.gz"#), "text/javascript"),
+("index-f136c4aa496c3185_bg.wasm", include_bytes!(r#"..\resource\asset/index-f136c4aa496c3185_bg.wasm.gz"#), "application/wasm"),
 ("index.css", include_bytes!(r#"..\resource\asset/index.css.gz"#), "text/css"),
 ("logo.png", include_bytes!(r#"..\resource\asset/logo.png.gz"#), ""),
 ("regular.min-a0c258fb7c5f655d.css", include_bytes!(r#"..\resource\asset/regular.min-a0c258fb7c5f655d.css.gz"#), "text/css"),
-("snippets/blog-frontend-3607fd663da6040e/asset/editor.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-3607fd663da6040e\asset\editor.js.gz"#), "text/javascript"),
+("snippets/blog-frontend-0b390d584a933dd3/asset/editor.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-0b390d584a933dd3\asset\editor.js.gz"#), "text/javascript"),
+("snippets/blog-frontend-0b390d584a933dd3/asset/show.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-0b390d584a933dd3\asset\show.js.gz"#), "text/javascript"),
 ("snippets/blog-frontend-bec5c12e8e7a4cdd/asset/editor.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-bec5c12e8e7a4cdd\asset\editor.js.gz"#), "text/javascript"),
 ("snippets/blog-frontend-bec5c12e8e7a4cdd/asset/show.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-bec5c12e8e7a4cdd\asset\show.js.gz"#), "text/javascript"),
 ("solid.min-70c2e5caa950974d.css", include_bytes!(r#"..\resource\asset/solid.min-70c2e5caa950974d.css.gz"#), "text/css"),
