+++ common/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "blog-common"
-version = "0.3.0"
+version = "0.5.0"
 authors = ["Songday"]
 edition = "2021"
