+++ frontend/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "blog-frontend"
-version = "0.3.0"
+version = "0.5.0"
 authors = ["Songday"]
 edition = "2021"
 
@@ -21,13 +21,13 @@ wasm-bindgen = "0.2"
 console_error_panic_hook = { version = "0.1", optional = true }
 
 blog-common = { path = "../common" }
+fastrand = "1.7"
 gloo = "0.6"
 gloo-file = "0.2"
 js-sys = "0.3"
 thiserror = "1.0"
 parking_lot = { version = "0.11", features = ["wasm-bindgen"]}
 lazy_static = "1.4"
-rand = "0.8"
 reqwasm = "0.3"
 serde = "1.0"
 serde_json = "1.0"
